1.修改日志的配置文件 2.修改后台返回的数据，去掉敏感数据。

+++ src/main/java/com/example/Galaxy/dao/BlogMapper.java
@@ -22,7 +22,7 @@ public interface BlogMapper {
 
     List<Blog> selectAll();
 
-    List<Blog> selectByUserId(@Param("userId") Integer userId);
+    List<Blog> selectByUserId(@Param("userId") Long userId);
 
     int addBlogOfLikeByBlogId(Blog blog);
 }
